requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
djangorestframework_simplejwt==4.4.0
PyJWT==v1.7.1
iteration-utilities==0.11.0
requests~=2.26.0
